pydaddy/requirements.txt
pyserial==3.5
digi-xbee==1.3.0
pyyaml==5.3.1
sqlalchemy==1.3.22
tabulate==0.8.7